보안 제외하고 2차 완성

+++ src/main/java/com/daesung/api/news/web/dto/NewsDto.java
@@ -25,7 +25,7 @@ public class NewsDto {
     private String content;
     private String newCompany;
     private String link;
-    private Long viewCnt;
+    private int viewCnt;
     private String language;
     private String thumbSummary;
